# pd112.rozetka.com/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Усіх із Різдвом",
        "description": "Demo my Project ",
        "contact": {
            "email": "admin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "3adf530c9aaf28030b7b676f7330305d",
                "responses": {
                    "200": {
                        "description": "List Categories."
                    }
                }
            }
        },
        "/api/categories/create": {
            "post": {
                "tags": [
                    "Category"
                ],
                "operationId": "1f00c6d1cc328c481361263c1add4c28",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "3d2944918e970d99b6fe9e61a0f237e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Categories."
                    },
                    "404": {
                        "description": "Wrong id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong Category Id has been sent. Pls try another one."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "operationId": "866c61c966fc32bd9ab53cc0856dfcfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішне видалення категорії"
                    },
                    "404": {
                        "description": "Категорії не знайдено"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                }
            }
        }
    }
}